Add table captions
xml does not support table captions. For this case, an additional text
element is added containing a table label and the caption.
To prevent the additional element from being displayed in the pdf,
the hidden attribute of the bulletin_element was renamed to "appear".
It contains now the output formats in which it should be processed
(appear).